pyproject.toml
[project]
name = "all-in-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.20.0",
    "chromadb>=0.4.0",
    "datasets>=2.14.0",
    "faiss-cpu>=1.7.0",
    "huggingface-hub>=0.33.4",
    "langchain==0.3.26",
    "langchain-community==0.3.27",
    "langchain-core==0.3.71",
    "langchain-deepseek==0.1.4",
    "langchain-experimental==0.3.4",
    "langchain-huggingface==0.3.1",
    "langchain-openai==0.3.28",
    "langchain-text-splitters==0.3.8",
    "llama-index==0.12.51",
    "llama-index-core==0.12.52",
    "llama-index-embeddings-huggingface==0.5.5",
    "llama-index-embeddings-openai==0.3.1",
    "llama-index-llms-deepseek==0.1.2",
    "llama-index-llms-openai==0.4.7",
    "markdown==3.8.2",
    "numpy>=1.24.0",
    "openai>=1.86.0,<2.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pymilvus==2.5.11",
    "pymilvus-model==0.3.2",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "sentence-transformers>=3.0.0",
    "tokenizers>=0.19.0",
    "torch==2.6.0",
    "torchaudio==2.6.0",
    "torchvision==0.21.0",
    "tqdm>=4.64.0",
    "transformers>=4.40.0",
    "unstructured[pdf]==0.18.11",
    "unstructured-client==0.41.0",
    "unstructured-inference==1.0.5",
    "unstructured-pytesseract==0.3.15",
    "opencv-python-headless==4.12.0.88",
    "einops==0.8.1",
    "ftfy==6.3.1",
    "llama-index-experimental==0.5.6",
    "openpyxl==3.1.5",
    "bilibili-api-python==17.3.0",
    "lark==1.2.2",
    "lazy-loader==0.4",
    "pyarrow==20.0.0",
]
